+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Changed
+
+- `warewulfd` overlay autobuild rebuilds overlays after node discovery. #1468
+
 ## v4.6.2, 2025-07-09
 
 ### Added